Add build action

+++ p4gpc.tinyadditions/Publish.ps1
@@ -1,5 +1,6 @@
 # Project Output Paths
-$modOutputPath = "TempBuild"
+$modOutputRootPath = "TempBuild"
+$modOutputPath = "$modOutputRootPath/p4gpc.tinyadditions"
 $solutionName = "p4gpc.tinyadditions.csproj"
 $publishName = "p4gpc.tinyadditions.zip"
 $publishDirectory = "Publish"
@@ -7,9 +8,13 @@ $publishDirectory = "Publish"
 [Environment]::CurrentDirectory = $PWD
 
 # Clean anything in existing Release directory.
-Remove-Item $modOutputPath -Recurse
-Remove-Item $publishDirectory -Recurse
-New-Item $modOutputPath -ItemType Directory
+if (Test-Path -Path $modOutputPath) {
+    Remove-Item $modOutputPath -Recurse
+}
+if (Test-Path -Path $publishDirectory) {
+    Remove-Item $publishDirectory -Recurse
+}
+New-Item $modOutputPath -ItemType Directory   
 New-Item $publishDirectory -ItemType Directory
 
 # Build
@@ -19,10 +24,18 @@ dotnet publish $solutionName -c Release -r win-x86 --self-contained false -o "$m
 dotnet publish $solutionName -c Release -r win-x64 --self-contained false -o "$modOutputPath/x64" /p:PublishReadyToRun=true
 
 # Remove Redundant Files
-Move-Item -Path "$modOutputPath/x86/ModConfig.json" -Destination "$modOutputPath/ModConfig.json"
-Move-Item -Path "$modOutputPath/x86/Preview.png" -Destination "$modOutputPath/Preview.png"
-Remove-Item "$modOutputPath/x64/Preview.png"
-Remove-Item "$modOutputPath/x64/ModConfig.json"
+if (Test-Path -Path "$modOutputPath/x86/ModConfig.json") {
+    Move-Item -Path "$modOutputPath/x86/ModConfig.json" -Destination "$modOutputPath/ModConfig.json"
+}
+if (Test-Path -Path "$modOutputPath/x86/Preview.png") {
+    Move-Item -Path "$modOutputPath/x86/Preview.png" -Destination "$modOutputPath/Preview.png"
+}
+if (Test-Path -Path "$modOutputPath/x64/Preview.png") {
+    Remove-Item "$modOutputPath/x64/Preview.png"
+}
+if (Test-Path -Path "$modOutputPath/x64/ModConfig.json") {
+    Remove-Item "$modOutputPath/x64/ModConfig.json"
+}
 
 # Cleanup Unnecessary Files
 Get-ChildItem $modOutputPath -Include *.exe -Recurse | Remove-Item -Force -Recurse
@@ -31,7 +44,7 @@ Get-ChildItem $modOutputPath -Include *.xml -Recurse | Remove-Item -Force -Recur
 
 # Compress
 Add-Type -A System.IO.Compression.FileSystem
-[IO.Compression.ZipFile]::CreateFromDirectory($modOutputPath, "$publishDirectory/$publishName")
+[IO.Compression.ZipFile]::CreateFromDirectory($modOutputRootPath, "$publishDirectory/$publishName")
 
 # Cleanup After Build
 Remove-Item $modOutputPath -Recurse
